fix(ci): split RCH Rust modules under size budgets
